game shows which tile you're hovering

+++ src/gui.zig
@@ -56,10 +56,10 @@ pub const Style = struct {
 };
 
 pub const ID = struct {
-    // lower 48 is what is updated by .newID()
-    // upper 16 is used by id.next(count)
+    // lower 32 is what is updated by .newID()
+    // upper 32 is used by id.next(count)
     id: u64 = 0,
-    pub fn next(id: ID, count: u16) ID {
+    pub fn next(id: ID, count: u32) ID {
         if (id.id > std.math.maxInt(u48)) @panic("next used on already next id");
         return .{ .id = id.id + std.math.maxInt(u48) * @as(u64, count) };
     }
